APIFEAST/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pandas==1.5.3
requests==2.28.2
mysql-connector-python==8.0.26
openpyxl
